Fix validation issue with container definitions
Turns out that `map(any)` means a map of any _single_ type.
So all values in the said map must have the same type.

Because of this we're changing the value to pure `any`.

Originally I was wanted to use the experimental optional feature.
But because it is experimental, it might change at any time.