Add variable to create listener rules
I've seen almost all use-cases of this and the predecessor modules
add a load balancer listener outside the module.

Seeing that we are creating the target group, this is a good addition
to the module.

Now users can just add `lb_listeners` to their code, and remove their
load balancer listener setup. Security groups to the load balancer in
question is also handled.

As an extra addition, this changes `task_deregistration_delay` to
`lb_[...]` to make it more consistent.